python-ocr/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
opencv-python
numpy
